+++ types/collections.go
@@ -32,11 +32,18 @@ var (
 	// IntValue represents a collections.ValueCodec to work with Int.
 	IntValue collcodec.ValueCodec[math.Int] = intValueCodec{}
 
+	// LegacyDecValue represents a collections.ValueCodec to work with LegacyDec.
+	LegacyDecValue collcodec.ValueCodec[math.LegacyDec] = legacyDecValueCodec{}
+
 	// TimeKey represents a collections.KeyCodec to work with time.Time
 	// Deprecated: exists only for state compatibility reasons, should not
 	// be used for new storage keys using time. Please use the time KeyCodec
 	// provided in the collections package.
 	TimeKey collcodec.KeyCodec[time.Time] = timeKeyCodec{}
+)
+
+const (
+	LegacyDec string = "math.LegacyDec"
 )
 
 type addressUnion interface {
@@ -164,6 +171,42 @@ func (i intValueCodec) Stringify(value math.Int) string {
 
 func (i intValueCodec) ValueType() string {
 	return "math.Int"
+}
+
+type legacyDecValueCodec struct{}
+
+func (i legacyDecValueCodec) Encode(value math.LegacyDec) ([]byte, error) {
+	return value.Marshal()
+}
+
+func (i legacyDecValueCodec) Decode(b []byte) (math.LegacyDec, error) {
+	v := new(math.LegacyDec)
+	err := v.Unmarshal(b)
+	if err != nil {
+		return math.LegacyDec{}, err
+	}
+	return *v, nil
+}
+
+func (i legacyDecValueCodec) EncodeJSON(value math.LegacyDec) ([]byte, error) {
+	return value.MarshalJSON()
+}
+
+func (i legacyDecValueCodec) DecodeJSON(b []byte) (math.LegacyDec, error) {
+	v := new(math.LegacyDec)
+	err := v.UnmarshalJSON(b)
+	if err != nil {
+		return math.LegacyDec{}, err
+	}
+	return *v, nil
+}
+
+func (i legacyDecValueCodec) Stringify(value math.LegacyDec) string {
+	return value.String()
+}
+
+func (i legacyDecValueCodec) ValueType() string {
+	return LegacyDec
 }
 
 type timeKeyCodec struct{}